Day after day

+++ src/day21/fr/Day21.kt
@@ -23,19 +23,19 @@ fun main(args: Array<String>) {
 
     val inLines = readInput("input21")
 
-    var setXYZon = mutableSetOf<Triple<Int, Int, Int>>()
+    var setXYZon = mutableSetOf<Triple<Long, Long, Long>>()
 
     var regex = "^(on|off)\\sx=(-?\\d+)..(-?\\d+),y=(-?\\d+)..(-?\\d+),z=(-?\\d+)..(-?\\d+)".toRegex()
 
     inLines.forEach {
         val match = regex.find(it)
         val vPos = match!!.groups[1]!!.value
-        val x1 = match!!.groups[2]!!.value.toInt()
-        val x2 = match!!.groups[3]!!.value.toInt()
-        val y1 = match!!.groups[4]!!.value.toInt()
-        val y2 = match!!.groups[5]!!.value.toInt()
-        val z1 = match!!.groups[6]!!.value.toInt()
-        val z2 = match!!.groups[7]!!.value.toInt()
+        val x1 = match!!.groups[2]!!.value.toLong()
+        val x2 = match!!.groups[3]!!.value.toLong()
+        val y1 = match!!.groups[4]!!.value.toLong()
+        val y2 = match!!.groups[5]!!.value.toLong()
+        val z1 = match!!.groups[6]!!.value.toLong()
+        val z2 = match!!.groups[7]!!.value.toLong()
 
         if (x1 in -50..50 && x2 in -50..50 && y1 in -50..50 && y2 in -50..50  && z1 in -50..50 && z2 in -50..50) {
             for (x in x1..x2) {
@@ -55,7 +55,6 @@ fun main(args: Array<String>) {
                 }
             }
         }
-
     }
 
     var cp = 0L
@@ -63,9 +62,7 @@ fun main(args: Array<String>) {
         cp++
     }
 
-
     println(cp)
-
 }
